backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
python-multipart>=0.0.18
aiofiles>=24.1.0
httpx>=0.28.0
